Rename did_auth
