+++ src/middleware/auth/did-auth.ts
