+++ src/middleware/authentication.ts
@@ -7,7 +7,7 @@ import { withLogto, handleAuthRoutes, LogtoExpressConfig } from '@logto/express'
 import { configLogToExpress } from '../types/constants.js'
 import { AccountAuthHandler } from './auth/account-auth.js'
 import { CredentialAuthHandler } from './auth/credential-auth.js'
-import { DidAuthHandler } from './auth/did_auth.js'
+import { DidAuthHandler } from './auth/did-auth.js'
 import { KeyAuthHandler } from './auth/key_auth.js'
 import { CredentialStatusAuthHandler } from './auth/credential-status-auth.js'
 import { AbstractAuthHandler } from './auth/base-auth.js'
